requirements.txt
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
comm==0.1.4
Cython==3.0.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.5
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
idna==3.4
imagej==0.3.1
imglyb==2.1.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.12.2
jedi==0.19.0
jgo==1.0.5
Jinja2==3.1.2
JPype1==1.4.1
json5==0.9.14
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-client==8.3.1
jupyter-core==5.3.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server==2.7.3
jupyter-server-terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-server==2.25.0
labeling==0.1.13
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.1
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook-shim==0.2.3
numpy==1.24.4
opencv-python==4.8.0.76
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil-resolve-name==1.3.10
platformdirs==3.10.0
plotly==5.17.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyimagej==1.4.1
pyjnius==1.5.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
scipy==1.10.1
scyjava==1.9.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.2.3
terminado==0.17.1
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
traitlets==5.10.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.3
xarray==2023.1.0
zipp==3.17.0
